% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areaRemoveZeroConflict.R
\name{areaRemoveZeroConflict}
\alias{areaRemoveZeroConflict}
\title{Remove Zero Conflicts}
\usage{
areaRemoveZeroConflict(data, value1, value2, observationFlag1, methodFlag1,
  missingObservationFlag = "M", missingMethodFlag = "u")
}
\arguments{
\item{data}{The data table object.}

\item{value1}{The column name of data corresponding to the first variable.}

\item{value2}{The column name of data corresponding to the second variable.}

\item{observationFlag1}{The column name of data containing the observation
flag for the first variable.}

\item{methodFlag1}{The column name of data containing the method flag for
the first variable.}

\item{missingObservationFlag}{The flag (character value) which should be
placed in the observation flag columns to signify a missing value.}

\item{missingMethodFlag}{The flag (character value) which should be placed
in the method flag columns to signify a missing value.}
}
\value{
No value is returned.  However, the object "data" which was passed
to this function is modified (some values are marked as missing if the have
conflicting zeroes).
}
\description{
The function examines two variables of a data.table object and helps 
to evaluate their feasible combinations. In particular if variable1 is
supposed to be greater then variable2 - such as areaSown and areaHarvested-
all the items characterized by variable1=0 and variable2>0 are marked as
missingObservationFlag_variable1 and missingMethodFlag_variable1
}

