% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/buildCPCHierarchy.R
\name{buildCPCHierarchy}
\alias{buildCPCHierarchy}
\title{Build CPC Hierarchy}
\usage{
buildCPCHierarchy(data, cpcItemVar, levels = 3)
}
\arguments{
\item{data}{}

\item{levels}{}

\item{cocItemVar}{}
}
\value{
modifies the data object
}
\description{
Function the build CPC Hierarchy
}

