% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensureNoConflictingAreaSownHarvested.R
\name{ensureNoConflictingAreaSownHarvested}
\alias{ensureNoConflictingAreaSownHarvested}
\title{Function to ensure that two value column does not contain conflicting zero
values.}
\usage{
ensureNoConflictingAreaSownHarvested(data, valueColumn1, valueColumn2,
  returnData = TRUE, normalised = TRUE,
  denormalisedKey = "measuredElement", getInvalidData = FALSE)
}
\arguments{
\item{data}{The hdata to be checked.}

\item{valueColumn1}{The first variable for comparison.}

\item{valueColumn2}{The second variable for comparison.}

\item{returnData}{logical, whether the data should be returned for pipe.}

\item{normalised}{logical, whether the data is normalised.}

\item{denormalisedKey}{optional, only required if the input data is not
normalised.It is the name of the key that denormalises the data.}

\item{getInvalidData}{logical, this will skip the test and extract the data
that is invalid.}
}
\value{
If getInvalidData is FALSE, then the data is returned when the test
    is cleared, otherwise an error. If getInvalidData is TRUE, then the
    subset of the data that is invalid is returned.
}
\description{
In the seed module for exaple, when area harvested is greater than zero, areaSown can not be
zero by definition (note that the opposite is allowed)
}

