% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getAreaData.R
\name{getAreaData}
\alias{getAreaData}
\title{Function for obtaining the area harvested/sown data}
\usage{
getAreaData(dataContext, areaSownElementCode = "5212",
  areaHarvestedElementCode = "5312", seedElementCode = "5525")
}
\arguments{
\item{dataContext}{The context for the data, as generated by the SWS.  This
object can be created via a call like swsContext.datasets[[1]] (assuming
the user is running this script on the SWS or after a call to
GetTestEnvironment).}

\item{areaSownElementCode}{The element code providing the dimension which
corresponds to the area sown variable in the database.}

\item{areaHarvestedElementCode}{The element code providing the dimension
which corresponds to the area harvested variable in the database.}

\item{seedElementCode}{The element code providing the dimension which
corresponds to the seed variable in the database.}
}
\value{
A data.table object containing the data queried from the database.
}
\description{
This function pulls the trade data from the database.  The main function
pulling the data is faosws::GetData, but additional steps are performed by
this function (such as setting up the appropriate pivot, adding variables
which are missing from the data as NA's, and setting data with missing
flags and 0 values to NA values).
}

