% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getCountryGeneralSeedRate.R
\name{getCountryGeneralSeedRate}
\alias{getCountryGeneralSeedRate}
\title{Get Country General Seed Rates}
\usage{
getCountryGeneralSeedRate()
}
\value{
The data.table object with specific seed rate data.
}
\description{
This function grabs data from the default_seed_rate table in the ess
schema in the database and applies a few simple operations to it (renames
columns, sets keys).  These default seed rates are provided at the
commodity level, as opposed to the commodity-country level in
getCountrySpecificSeedRate().
}
\details{
The seed rates that are pulled from the database are assumed to be official
values.  That is, the observation status flags of all of these observations
are set to "".
}

