% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getCountrySpecificSeedRate.R
\name{getCountrySpecificSeedRate}
\alias{getCountrySpecificSeedRate}
\title{Function to load the country specific rate data}
\usage{
getCountrySpecificSeedRate()
}
\value{
The data.table object with specific seed rate data.
}
\description{
This function grabs data from the specific_seed_rate table in the ess
schema in the database and applies a few simple operations to it (renames
columns, converters dimensions to characters, and sets all flags that
aren't "E" to "").  These rates are provided at the commodity-country level
as opposed to commodity level in getCountrySpecificSeedRate().
}
\details{
The seed rates that are pulled from the database are assumed to be official
values.  That is, the observation status flags of all of these observations
are set to "".
}

