% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOfficialSeedData.R
\name{getOfficialSeedData}
\alias{getOfficialSeedData}
\title{Get Official Seed Data}
\usage{
getOfficialSeedData()
}
\arguments{
\item{areaVar}{character. Name of area key in source dataset}
}
\description{
Get Official Seed Data
}

