% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSelectedSeedData.R
\name{getSelectedSeedData}
\alias{getSelectedSeedData}
\title{Get Selected Seed Data}
\usage{
getSelectedSeedData(dataContext)
}
\arguments{
\item{dataContext}{The DatasetKey, containing the relevant countries, years,
commodities and elements for pulling the seed data.}
}
\value{
The dataset with seed data.
}
\description{
This function pulls the seed data from the SWS.
}

