% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imputeAreaSown.R
\name{imputeAreaSown}
\alias{imputeAreaSown}
\title{Get the area harvested/sown data}
\usage{
imputeAreaSown(data, valueAreaSown = "Value_measuredElement_5212",
  valueAreaHarvested = "Value_measuredElement_5312",
  flagObsAreaSown = "flagObservationStatus_measuredElement_5212",
  flagObsAreaHarvested = "flagObservationStatus_measuredElement_5312",
  imputedFlag = "i")
}
\arguments{
\item{data}{The data.table object containing the seed data, usually as
created by getAreaData.}

\item{valueAreaSown}{The column name of data which contains the value of
the area sown variable.}

\item{valueAreaHarvested}{The column name of data which contains the value of
the area harvested variable.}

\item{flagObsAreaSown}{The column name of data which contains the
observation flag for the area sown variable.}

\item{flagObsAreaHarvested}{The column name of data which contains the
observation flag for the area harvested variable.}

\item{imputedFlag}{Currently unused: Michael?}
}
\value{
No object is returned, instead the underlying data.table object is
modified.
}
\description{
Imputation is currently quite simple: if the area sown variable has no
data, then area sown is assumed to be equal to area harvested.  If some
values exist and some are missing, than an average ratio is computed of
area sown to area harvested and this ratio is applied to all missing
values.
}

