% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imputeSeed.R
\name{imputeSeed}
\alias{imputeSeed}
\title{Function to impute the seed.}
\usage{
imputeSeed(data, seedValue = "Value_measuredElement_5525",
  seedMethodFlag = "flagMethod_measuredElement_5525",
  seedObsFlag = "flagObservationStatus_measuredElement_5525",
  areaSownValue = "Value_measuredElement_5212",
  areaSownObsFlag = "flagObservationStatus_measuredElement_5212",
  seedRateValue = "Value_areaSownRatio",
  seedRateFlag = "flagObservationStatus_areaSownRatio", imputedFlag = "i",
  byKey = key(data))
}
\arguments{
\item{data}{The data.table object containing the seed data, typically as
created by getAreaData.}

\item{seedValue}{The column name of data which contains the value of the
seed variable.}

\item{seedMethodFlag}{The column name of data which contains the method
flag of the seed variable.}

\item{seedObsFlag}{The column name of data which contains the
observation flag of the seed variable.}

\item{areaSownValue}{The column name of data which contains the value of
the area sown variable.}

\item{areaSownObsFlag}{The column name of data which contains the
observation flag of the area sown variable.}

\item{seedRateValue}{The column name of data which contains the seed rate
variable.}

\item{seedRateFlag}{The column name of data which contains the seed rate
flag.}

\item{imputedFlag}{When the seed value is imputed, what character value
should be assigned to the seedMethodFlag?}

\item{byKey}{A character vector of the column name(s) of data which should
be treated as keys for performing the imputation.}
}
\value{
No value is returned, instead data is modified by filling in
missing values with imputed estimates.
}
\description{
First, an estimate for the seed value at time t is generated by multiplying
the area sown at time t-1 by the seed rate (divided by 1,000) at time t.
Then, the estimate is inserted into the seed data.table if the seed value
at time t is currently missing.  The observation flag is updated by
aggregating the flags of the seed rate and area sown, and the method flag
is updated with the value of imputedFlag (passed as an argument).  Lastly,
the seedRateValue and seedRateFlag columns are deleted from the data.table.
}

