% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imputeSeedDomain.R
\name{imputeSeedDomain}
\alias{imputeSeedDomain}
\title{Impute Seed Domain}
\usage{
imputeSeedDomain(data, imputationParameters, byKey = NULL)
}
\arguments{
\item{data}{The data.table object containing the seed data.}

\item{imputationParameters}{See
?faoswsImputation::defaultImputationParameters.  This is a list of
arguments specifying how imputation should be done, and it pertains to
imputation of the area sown variable.  If imputation should be done by a
simple mean, set imputationParameters to NULL.}

\item{byKey}{In the case where imputationParameters is NULL, imputation is
done by computing a mean within each group defined by byKey and using
that mean for each missing value.  byKey = NULL, the default, uses a global
mean for each commodity.  If imputationParameters is not NULL, this is
ignored.}
}
\value{
No value is returned, but the data.table object is modifed: seed is
imputed and additional columns are returned.
}
\description{
This function is a wrapper function which calls all the individual
functions of this package.  First, values where the area sown is 0 and the
area harvested is non-zero (or vice-versa) are removed.  Next, area sown is
imputed, and then country specific and general seed rates are appended to
the data.table.  Lastly, the seed is imputed.
}

