% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeAllSeedData.R
\name{mergeAllSeedData}
\alias{mergeAllSeedData}
\title{Merge All Seed Data}
\usage{
mergeAllSeedData(seedData, ...)
}
\arguments{
\item{seedData}{The data.table object containing the seed data.}

\item{...}{Additional data.table objects to be merged with seedData.}
}
\value{
A data.table containing the result of the merges.
}
\description{
This function merges the seed data data.table object with any additional
datasets, using the intersection of the columns as the merging key(s).
}

