% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeCarryForward.R
\name{removeCarryForward}
\alias{removeCarryForward}
\title{Remove Carry Forward}
\usage{
removeCarryForward(data, variable)
}
\arguments{
\item{data}{The data.table containing the seed data.  It should have columns
geographicAreaM49 and measuredItemCPC.}

\item{variable}{The column name of data which contains the values that
should be checked.}
}
\value{
The data.table with the "carried forward" values removed.
}
\description{
In the old SWS, estimates were occassionally "carried forward", i.e. the 
estimate for the current year was just a copy of the estimate for the 
previous year.  This function removes such estimates from the dataset.
}
\details{
Note: I'm not sure if this function is still relevant, as all "trended" 
(i.e. carried forward) values from the old system were NOT copied to the new
system.  However, before removing this function, we should confirm that it 
is not needed.
}

