% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/saveSeedData.R
\name{saveSeedData}
\alias{saveSeedData}
\title{Save Seed Data}
\usage{
saveSeedData(data)
}
\arguments{
\item{data}{The data.table object containing the seed data to be written to
the database.}
}
\value{
No R objects are returned, as this function's purpose is solely to
write to the database.
}
\description{
This function takes the seed data dataset and saves it back to the
database.
}

