% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCPCHierarchy.R
\name{buildCPCHierarchy}
\alias{buildCPCHierarchy}
\title{Build CPC Hierarchy}
\usage{
buildCPCHierarchy(data, cpcItemVar, levels = 3)
}
\arguments{
\item{data}{A data.table containing a column with CPC codes.}

\item{cpcItemVar}{Character.  The column name of data which corresponds to 
the item/CPC code (for example, "measuredItemCPC").}

\item{levels}{How many CPC levels should be constructed?}
}
\value{
Modifies the data object by adding cpcLvl columns (there will be
  "levels" columns added).
}
\description{
This function creates a CPC hierarchy by taking the substring of the CPC 
code, up to "levels" number of digits.  For example, wheat (0111) would be 
in the group 011 if levels=3 and group 0 if levels=1.  Note that
}

