% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fillCountrySpecificSeedRate.R
\name{fillCountrySpecificSeedRate}
\alias{fillCountrySpecificSeedRate}
\title{Fill country specific seed rates}
\usage{
fillCountrySpecificSeedRate(data, countryVariable = "geographicAreaM49",
  commodityVariable = "measuredItemCPC",
  countrySpecificData = getCountrySpecificSeedRate())
}
\arguments{
\item{data}{The seed data.table, typically as produced by a call to
getAreaData.}

\item{countryVariable}{The column name of data that specifies the country
code variable.  This is needed to join with the table from the database.}

\item{commodityVariable}{The column name of data that specifies the
commodity code variable.  This is needed to join with the table from the
database.}

\item{countrySpecificData}{A data.table with data describing seed rates for
each country.}
}
\value{
No value is returned.  Instead, seed rates (from
countrySpecificData) are appended onto data.
}
\description{
Fill country specific seed rates
}

