% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imputeAreaSown.R
\name{imputeAreaSown}
\alias{imputeAreaSown}
\title{Get the area harvested/sown data}
\usage{
imputeAreaSown(data, codeAreaHarvested = "5312", codeAreaSown = "5212",
  imputedObsFlag = "I", imputedMethodFlag = "e", byKey = NULL,
  imputationParameters = NULL)
}
\arguments{
\item{data}{The data.table object containing the seed data, usually as
created by getAreaData.}

\item{codeAreaHarvested}{The element code for the area harvested variable.}

\item{codeAreaSown}{The element code for the area sown variable.}

\item{imputedObsFlag}{The value to be assigned to the observation status
flag for imputed observations.}

\item{imputedMethodFlag}{The value to be assigned to the method flag for
imputed observations.}

\item{byKey}{Only used if imputationParameters is NULL (if
imputationParameters is not NULL, ensemble imputation is performed and the
byKey variable should be specified in this list).  This value
specifies how the mean seed to harvest ratio should be computed: by
country, globally, or in some other way.  Should be a column name of data,
or NULL (in which case a global mean is used).}

\item{imputationParameters}{A vector containing the parameters to use in
the imputation of areaSown variable.  If NULL, imputation is done by
computing the mean ratio of area harvested to area sown and applying that
to all cases.}
}
\value{
No object is returned, instead the underlying data.table object is
modified.
}
\description{
Imputation is currently quite simple: if the area sown variable has no
data, then area sown is assumed to be equal to area harvested.  If some
values exist and some are missing, than an average ratio is computed of
area sown to area harvested and this ratio is applied to all missing
values.
}

